src/app/photo-list/photo-list.component.ts
import { Component, OnInit } from '@angular/core';
import { PhotoService, Photo } from '../photo.service';

@Component({
  selector: 'app-photo-list',
  templateUrl: './photo-list.component.html',
  styleUrls: ['./photo-list.component.css']
})
export class PhotoListComponent implements OnInit {
  photos: Photo[] = [];

  constructor(private photoService: PhotoService) { }

  ngOnInit(): void {
    this.photoService.getPhotos().subscribe((data: Photo[]) => {
      this.photos = data;
    });
  }
}
